Add function for calculating a list of Basic operations
Basic operations are U, CX, Measure and Reset. In other words all
operations except Custom. This is useful because Custom operations are a
bit more complex to handle since they can refer to other gates. Using
this method it should be easy to just iterate through the different
operations and apply them in a simulator.